Add VM instance cfg, include it in digest generated by VM instance RoT
The JSON file in `test-data` includes a UUID & a digest intended to
represent the digest of the instance image. The `AttestMock` rolls this
data into the digest passed down to the `RotType::OxideHardware` where
it's included in the resultant attestation. The test that reconstructs
the message signed by the `RotType::OxideHardware` uses this config as a
reference manifest.